requirements.txt
absl-py==0.12.0
aniso8601==9.0.1
appnope==0.1.2
asgiref==3.3.4
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
decorator==5.0.9
Django==3.2.4
djangorestframework==3.12.4
facenet-face-recognition==0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
imutils==0.5.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpickle==2.0.0
jupyter-client==6.1.12
jupyter-core==4.7.1
Keras==2.4.3
keras-facenet==0.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mtcnn==0.1.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.18
sqlparse==0.4.1
tensorflow-cpu==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1



 source venv/bin/activate     
 python consolapp.py    
  python api.py 
  pip install mysql-connector-python 
pip install opencv-python==4.5.2.54
pip install sklearn==0.0   
pip install scipy==1.6.3
pip install flask_restful          
pip install -U flask-cors     
 pip install flask                  
pip install flask-sqlalchemy
pip install matplotlib             
pip install jsonpickle      
 pip install -U git+https://github.com/mortuzahossain/face-recognition.git                                     



